requirements.txt
flask
pyyaml
crewai
crewai-tools
